Add vercel config and remove firebase files
