--- .firebaserc
@@ -1,5 +1,0 @@
-{
-  "projects": {
-    "default": "chordwizard-2f21e"
-  }
-}
