--- firebase.json
@@ -1,12 +1,0 @@
-{
-  "hosting": {
-    "public": "public",
-    "ignore": [
-      "firebase.json",
-      "**/.*",
-      "**/node_modules/**",
-      "schematics/**",
-      "temp/**"
-    ]
-  }
-}
