+++ vercel.json
@@ -1,0 +1,14 @@
+{
+    "builds": [
+        {
+            "src": "index.js",
+            "use": "@vercel/node"
+        }
+    ],
+    "routes": [
+        {
+          "src": "/(.*)",
+          "dest": "index.js"
+        }
+    ]
+  }
